@@ -1,9 +1,8 @@
 package com.example.smilyrating;
 
-import android.graphics.Typeface;
 import android.os.Bundle;
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 
 import com.hsalf.smilerating.BaseRating;
