@@ -9,22 +9,12 @@ public class Okay extends Smiley {
 
     public Okay() {
         super(-135, 360);
-
         createStraightSmile(new Point(CENTER_SMILE, MOUTH_CENTER_Y), (CENTER_SMILE * 0.1f), 350f, (CENTER_SMILE * 0.9f));
+        setup(
+                getClass().getSimpleName(),
+                Color.parseColor("#f2dd68"),
+                Color.parseColor("#353431")
+        );
     }
 
-    @Override
-    public String getName() {
-        return getClass().getSimpleName();
-    }
-
-    @Override
-    public int getFaceColor() {
-        return Color.parseColor("#f2dd68");
-    }
-
-    @Override
-    public int getDrawingColor() {
-        return Color.parseColor("#353431");
-    }
 }
